+++ core/base/CoreModuleInterface.php
@@ -1,0 +1,8 @@
+<?php
+
+namespace luya\base;
+
+interface CoreModuleInterface
+{
+    
+}