+++ modules/gallery/src/frontend/Module.php
@@ -2,22 +2,19 @@
 
 namespace luya\gallery\frontend;
 
+use luya\base\CoreModuleInterface;
+
 /**
  * Gallery Module
  * 
  * @author Basil Suter <basil@nadar.io>
  */
-class Module extends \luya\base\Module
+class Module extends \luya\base\Module implements CoreModuleInterface
 {
     /**
      * @var boolean This module does not have view files, so the view are looked up in the application folder.
      */
     public $useAppViewPath = true;
-    
-    /**
-     * @var boolean This module will be hidden from several selections.
-     */
-    public $isCoreModule = true;
 
     public $urlRules = [
         ['pattern' => 'gallery/kategorie/<catId:\d+>/<title:[a-zA-Z0-9\-]+>/', 'route' => 'gallery/alben/index'],