+++ modules/news/src/admin/Module.php
@@ -3,11 +3,10 @@
 namespace luya\news\admin;
 
 use Yii;
+use luya\base\CoreModuleInterface;
 
-class Module extends \luya\admin\base\Module
+class Module extends \luya\admin\base\Module implements CoreModuleInterface
 {
-    public $isCoreModule = true;
-
     public $apis = [
         'api-news-article' => 'luya\news\admin\apis\ArticleController',
         'api-news-tag' => 'luya\news\admin\apis\TagController',